pyproject.toml
[project]
name = "cursed-controls"
version = "0.1.0"
description = "Map any evdev input device to a virtual Xbox 360 wireless receiver over USB OTG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "evdev>=1.9.2",
    "pyyaml>=6.0.3",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
]

[project.scripts]
cursed-controls = "cursed_controls.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["cursed_controls*"]
